fix availableForWrite() signature

+++ cores/arduino/Adafruit_TinyUSB_Core/Adafruit_USBD_CDC.cpp
@@ -126,7 +126,7 @@ size_t Adafruit_USBD_CDC::write(const uint8_t *buffer, size_t size)
   return size - remain;
 }
 
-size_t Adafruit_USBD_CDC::availableForWrite(void)
+int Adafruit_USBD_CDC::availableForWrite(void)
 {
   return tud_cdc_write_available();
 }
